+++ instrumentation/akka/akka-actor-fork-join-2.5/javaagent/src/test/java/io/opentelemetry/instrumentation/akkaactor/AkkaAsyncChild.java
@@ -8,7 +8,7 @@ package io.opentelemetry.instrumentation.akkaactor;
 import akka.dispatch.forkjoin.ForkJoinTask;
 import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.api.trace.Tracer;
-import io.opentelemetry.javaagent.instrumentation.javaconcurrent.TestTask;
+import io.opentelemetry.javaagent.instrumentation.executors.TestTask;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicBoolean;
