+++ instrumentation/executors/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/executors/CallableInstrumentation.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.javaconcurrent;
+package io.opentelemetry.javaagent.instrumentation.executors;
 
 import static io.opentelemetry.javaagent.extension.matcher.AgentElementMatchers.implementsInterface;
 import static net.bytebuddy.matcher.ElementMatchers.isPublic;
@@ -16,32 +16,33 @@ import io.opentelemetry.javaagent.bootstrap.concurrent.PropagatedContext;
 import io.opentelemetry.javaagent.bootstrap.concurrent.TaskAdviceHelper;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
+import java.util.concurrent.Callable;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 
-public class RunnableInstrumentation implements TypeInstrumentation {
+public class CallableInstrumentation implements TypeInstrumentation {
 
   @Override
   public ElementMatcher<TypeDescription> typeMatcher() {
-    return implementsInterface(named(Runnable.class.getName()));
+    return implementsInterface(named(Callable.class.getName()));
   }
 
   @Override
   public void transform(TypeTransformer transformer) {
     transformer.applyAdviceToMethod(
-        named("run").and(takesArguments(0)).and(isPublic()),
-        RunnableInstrumentation.class.getName() + "$RunnableAdvice");
+        named("call").and(takesArguments(0)).and(isPublic()),
+        CallableInstrumentation.class.getName() + "$CallableAdvice");
   }
 
   @SuppressWarnings("unused")
-  public static class RunnableAdvice {
+  public static class CallableAdvice {
 
     @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static Scope enter(@Advice.This Runnable thiz) {
-      VirtualField<Runnable, PropagatedContext> virtualField =
-          VirtualField.find(Runnable.class, PropagatedContext.class);
-      return TaskAdviceHelper.makePropagatedContextCurrent(virtualField, thiz);
+    public static Scope enter(@Advice.This Callable<?> task) {
+      VirtualField<Callable<?>, PropagatedContext> virtualField =
+          VirtualField.find(Callable.class, PropagatedContext.class);
+      return TaskAdviceHelper.makePropagatedContextCurrent(virtualField, task);
     }
 
     @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
