+++ instrumentation/executors/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/executors/CompletableFutureTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.javaconcurrent;
+package io.opentelemetry.javaagent.instrumentation.executors;
 
 import static org.assertj.core.api.Assertions.assertThat;
