+++ instrumentation/executors/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/executors/StreamTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.javaconcurrent;
+package io.opentelemetry.javaagent.instrumentation.executors;
 
 import io.opentelemetry.instrumentation.testing.junit.AgentInstrumentationExtension;
 import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
