+++ instrumentation/executors/testing/src/main/java/io/opentelemetry/javaagent/instrumentation/executors/TestTask.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.javaconcurrent;
+package io.opentelemetry.javaagent.instrumentation.executors;
 
 import java.util.concurrent.Callable;
